requirements.txt
aiogram>=3.4
SQLAlchemy>=2.0
aiosqlite>=0.20
aiocryptopay>=0.4
python-dotenv>=1.0
